pyproject.toml
[project]
name = "uncluttered"
version = "0.1.0"
description = "AI-powered recipe extraction. No ads, no stories, just ingredients and steps."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["recipe", "cli", "ai", "cooking"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "tavily-python>=0.3.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.3.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.40.0"]
ollama = ["openai>=1.0.0"]
all = ["openai>=1.0.0", "anthropic>=0.40.0"]

[project.scripts]
uncluttered = "uncluttered.cli.main:app"

[project.urls]
Homepage = "https://uncluttered.recipes"
Repository = "https://github.com/brendaninnis/uncluttered-cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
